requirements.txt
Django>=5.0,<5.2
PyPDF2>=3.0.0
python-dateutil>=2.9.0
